Added MainItemActivity
Displays information of a single activity (image, description,
location). Location is currently a TextView. When clicked, sends intent
to display pin on room face of the room. See TODO to implement. Updates
view count and last access when an item is viewed.

Updated MainActivity to allow for items to be clicked on to display the
item's information.