Finished Item Search
Search is now finished. Accessed from the main activity.

Displays results in alphabetical order (case insensitive) using a
ListView.
Items are pulled if it contains a substring that matches the query (case
insensitive). If a search has no matching item names, it displays
"Nothing found." in place of the ListView.